Uniform `base-/extended-format device-context(s)` use
"base/extended format" was used in the text, while in the description of
the device-context (section 2.1.2) "base-/extended-format" was used.
Modified it to match section 2.1.2 syntax (with the hyphen).

Also made device-context plural as device-contexts.
While both formats are correct used the version with "s" at the end as
it seems to be the way it is written in most of the document.