Fix DDT and PDT ditaa diagrams
The DDT and PDT diagrams (Figures 10, 11 and 23) were badly rendered.
The issue is the usage of the `-` character both for drawing and the
actual text (e.g. the `-` in 9-bit).
The solution is to use a different unicode character for the text
sections that looks similar.
In this commit we used the unicode character with code U+2212 instead of
the '-' character of your keyboard (which has the unicode code U+002B).
Other potential unicode characters might be found in the following
links:
- https://www.compart.com/en/unicode/category/Pd
- https://www.compart.com/en/unicode/bidiclass/ES