Add iommu_t struct to iommu_registers.h
Instead of using global variables to hold the state of the IOMMU, all of the
state including caches has been added to a new struct called iommu_t which is
defined in iommu_registers.h.

This change makes it slightly more tedious to use the API because almost
all functions now take an extra parameter which is a pointer to the
relevant iommu_t instance, but it has the benefit that you can
instantiate multiple iommu_t's, thus making it possible to model systems
which have multiple IOMMUs.

It also makes it easier to save/restore the state of the IOMMU and makes
it easier to find the variables which constitute the state of the IOMMU.

Signed-off-by: Paul Kennedy <pkennedy@tenstorrent.com>